# [SPARK-32072][CORE][TESTS] Fix table formatting with benchmark results