--- a/sql/core/benchmarks/MakeDateTimeBenchmark-jdk11-results.txt
+++ b/sql/core/benchmarks/MakeDateTimeBenchmark-jdk11-results.txt
@@ -1,22 +1,22 @@
-OpenJDK 64-Bit Server VM 11.0.5+10-post-Ubuntu-0ubuntu1.118.04 on Linux 4.15.0-1044-aws
+OpenJDK 64-Bit Server VM 11.0.7+10-post-Ubuntu-2ubuntu218.04 on Linux 4.15.0-1063-aws
 Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
 make_date():                              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-prepare make_date()                                3204           3323         139         31.2          32.0       1.0X
-make_date(2019, 9, 16)                             2529           2604         126         39.5          25.3       1.3X
-make_date(*, *, *)                                 5102           5113          10         19.6          51.0       0.6X
+prepare make_date()                                3214           3373         240         31.1          32.1       1.0X
+make_date(2019, 9, 16)                             2366           2513         225         42.3          23.7       1.4X
+make_date(*, *, *)                                 4654           4733         102         21.5          46.5       0.7X
 
-OpenJDK 64-Bit Server VM 11.0.5+10-post-Ubuntu-0ubuntu1.118.04 on Linux 4.15.0-1044-aws
+OpenJDK 64-Bit Server VM 11.0.7+10-post-Ubuntu-2ubuntu218.04 on Linux 4.15.0-1063-aws
 Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
-make_timestamp():                         Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
-------------------------------------------------------------------------------------------------------------------------
-prepare make_timestamp()                           3484           3513          28          0.3        3484.3       1.0X
-make_timestamp(2019, 1, 2, 3, 4, 50.123456)            112            131          17          9.0         111.5      31.2X
-make_timestamp(2019, 1, 2, 3, 4, 60.000000)             93            102          10         10.8          92.8      37.6X
-make_timestamp(2019, 12, 31, 23, 59, 60.00)             85             88           4         11.8          84.8      41.1X
-make_timestamp(*, *, *, 3, 4, 50.123456)            303            308           8          3.3         302.8      11.5X
-make_timestamp(*, *, *, *, *, 0)                    303            307           3          3.3         302.8      11.5X
-make_timestamp(*, *, *, *, *, 60.0)                 289            297           8          3.5         289.1      12.1X
-make_timestamp(2019, 1, 2, *, *, *)                3576           3585          11          0.3        3576.4       1.0X
-make_timestamp(*, *, *, *, *, *)                   3610           3618          12          0.3        3610.0       1.0X
+make_timestamp():                            Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
+---------------------------------------------------------------------------------------------------------------------------
+prepare make_timestamp()                              3420           3459          51          0.3        3420.2       1.0X
+make_timestamp(2019, 1, 2, 3, 4, 50.123456)             86             95          11         11.6          86.5      39.5X
+make_timestamp(2019, 1, 2, 3, 4, 60.000000)             82             85           2         12.2          82.1      41.7X
+make_timestamp(2019, 12, 31, 23, 59, 60.00)             76             82          10         13.2          75.8      45.1X
+make_timestamp(*, *, *, 3, 4, 50.123456)               310            313           3          3.2         310.0      11.0X
+make_timestamp(*, *, *, *, *, 0)                       291            298           8          3.4         290.8      11.8X
+make_timestamp(*, *, *, *, *, 60.0)                    290            292           3          3.4         290.5      11.8X
+make_timestamp(2019, 1, 2, *, *, *)                   3555           3566           9          0.3        3555.0       1.0X
+make_timestamp(*, *, *, *, *, *)                      3594           3605          14          0.3        3593.8       1.0X